Rogue Stealth Ranks - Update spell_restorations.sql
The new stealth ranks would not show up at rogue trainers.
changing the ReqSpell to 0 fixes this

I also added the new stealth spells to spell_ranks

To make the new spells show up in the Subtlety tab after learning them these changes need to be made in SkillLineAbility.dbc
both for the server and in Patch-V.mpq
https://i.gyazo.com/bdd1623d447be10535fca16fe1b020e9.png

because this was never working before and only rank 1 was showing up at rogue trainers, rank 1 was set to the values of rank 4. 
To undo this changes need to be made in spells.dbc for both the server and the client in Patch-V.mpq
the speed penalty for rank 1 should be set to -50%

I am still unable to make the next rank overwrite the previous rank. the player ends up with 4 stealth spells in both the spell book and above the skill bars.

I've been comparing the stealth skill with the paladin devotion skill to figure out how to fix this but I can't find the solution.